+++ nas/ies/_5gmmCause.hpp
@@ -81,7 +81,7 @@ class _5gmmCause : public Type3NasIe {
   ~_5gmmCause();
 
   static std::string GetIeName() { return k5gmmCauseIeName; }
-  uint16_t GetIeLength() const override;
+  uint32_t GetIeLength() const override;
 
   int Encode(uint8_t* buf, int len);
   int Decode(uint8_t* buf, int len, bool is_iei);
